# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sports Match Prediction API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /predictions:
    get:
      operationId: listMyPredictions
      tags: [predictions]
      summary: Get my predictions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, won, lost, all]
            default: all
      responses:
        "200":
          description: My predictions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prediction"
    post:
      operationId: createPrediction
      tags: [predictions]
      summary: Submit a new prediction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredictionInput"
      responses:
        "201":
          description: Prediction submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prediction"
        "400":
          description: Invalid input
  /predictions/{id}:
    patch:
      operationId: updatePrediction
      tags: [predictions]
      summary: Edit a prediction (only before deadline)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredictionUpdate"
      responses:
        "200":
          description: Prediction updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prediction"
        "400":
          description: Deadline passed or invalid input
        "404":
          description: Not found
    delete:
      operationId: deletePrediction
      tags: [predictions]
      summary: Delete a prediction (only before deadline)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "400":
          description: Deadline passed
        "404":
          description: Not found
...
components:
...
  schemas:
...
    Team:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        logo:
          type: string
        country:
          type: string
      required: [id, name, logo, country]
    League:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        logo:
          type: string
        country:
          type: string
        season:
          type: string
      required: [id, name, logo, country, season]
    Match:
      type: object
      properties:
        id:
          type: string
        homeTeam:
          $ref: "#/components/schemas/Team"
        awayTeam:
          $ref: "#/components/schemas/Team"
        league:
          $ref: "#/components/schemas/League"
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [upcoming, live, finished, postponed]
        homeScore:
          type: ["integer", "null"]
        awayScore:
          type: ["integer", "null"]
        minute:
          type: ["integer", "null"]
        venue:
          type: string
      required: [id, homeTeam, awayTeam, league, scheduledAt, status, venue]
...
    Prediction:
      type: object
      properties:
        id:
          type: string
        match:
          $ref: "#/components/schemas/Match"
        homeScorePrediction:
          type: ["integer", "null"]
        awayScorePrediction:
          type: ["integer", "null"]
        firstGoalscorer:
          type: ["string", "null"]
        manOfMatch:
          type: ["string", "null"]
        totalGoalsPrediction:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, won, lost, partial]
        pointsEarned:
          type: ["integer", "null"]
        submittedAt:
          type: string
          format: date-time
      required: [id, match, status, submittedAt]
    PredictionInput:
      type: object
      properties:
        matchId:
          type: string
        homeScorePrediction:
          type: ["integer", "null"]
        awayScorePrediction:
          type: ["integer", "null"]
        firstGoalscorer:
          type: ["string", "null"]
        manOfMatch:
          type: ["string", "null"]
        totalGoalsPrediction:
          type: ["integer", "null"]
      required: [matchId]
...
    PredictionUpdate:
      type: object
      properties:
        homeScorePrediction:
          type: ["integer", "null"]
        awayScorePrediction:
          type: ["integer", "null"]
        firstGoalscorer:
          type: ["string", "null"]
        manOfMatch:
          type: ["string", "null"]
        totalGoalsPrediction:
          type: ["integer", "null"]
...