# lib/api-zod/src/generated/types/predictionUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Sports Match Prediction API
 * OpenAPI spec version: 0.1.0
 */

export interface PredictionUpdate {
  /** @nullable */
  homeScorePrediction?: number | null;
  /** @nullable */
  awayScorePrediction?: number | null;
  /** @nullable */
  firstGoalscorer?: string | null;
  /** @nullable */
  manOfMatch?: string | null;
  /** @nullable */
  totalGoalsPrediction?: number | null;
}